Deprecate ClientListener dispose and isDisposed
Disposal of a client listener was introduced in order to destroy the
remote object but this feature was not yet effective.
For an SWT Listener, a dispose method is unexpected. If necessary, we
should to destroy the remote object internally when the listener is no
longer bound to any widget.