Remove Script class
The Script class was created in an attempt to have a simple way to share
script code for ClientListener without transporting it multiple times
within one session, while also allow using the code in other ways.
However, this is more easily achieved registering the code as a resource
and loading it either by JavaScriptLoader or a (later on) a new
ClientListener API. This also allows caching the code.